implementación y optimizacion de trigger Start LavaRise y StopLavaRise y corrección de plataformas desvanecentes
The code for the lava triggers was implemented so that the lava starts to rise with a StartLavaRise trigger and so that it stops with a StopLavaRise trigger. In addition, the lava is reset if the player dies. The Lavacannon code was optimized so that it is now a single code, the same for the lava projectile code.