Chamber dash (#19)
* Implementation of lava ball and dash test scene

Implementation of lava ball that goes up and down between platforms,
also implemented the dash test scene where the lava level is being built to test the mechanics of the dash in the level

* Implementation of lava pipes with horizontal projectiles

Implementation of lava pipes with horizontal projectiles cannons were implemented that shoot lava balls horizontally from left and right in the part of the level where the lava rises

* Implementation of the rising lava

simple implementation of the rising lava along the level logo

* implementación y optimizacion de trigger Start LavaRise y StopLavaRise y corrección de plataformas desvanecentes

The code for the lava triggers was implemented so that the lava starts to rise with a StartLavaRise trigger and so that it stops with a StopLavaRise trigger. In addition, the lava is reset if the player dies. The Lavacannon code was optimized so that it is now a single code, the same for the lava projectile code.

* Preventing tilemap bug when merging with main

* Lava Trigger and Lava Rise script fixes

* Lavacanon Tab script fix in class